v0.4.6: - fixed SafeModelCheckPoint when save_best_only=True - hard_sample_mining: average metrics per probability

+++ dataset_iterator/keras_callbacks.py
@@ -17,35 +17,49 @@ class SafeModelCheckpoint(ModelCheckpoint):
         self._alternative_path = False
 
     def _save_model(self, epoch, batch, logs):
-        for i in range(self.n_retry):
-            if i>1:
-                print(f"save_model: start of for loop: {i + 1}/{self.n_retry}")
-            try:
-                path = self._get_file_path(epoch, batch, logs)
-                tmp_path = self._tmp_path(path)
-                #print(f"save_model: try {i + 1}/{self.n_retry}, path: {path} temp path: {tmp_path}", flush=True)
-                if os.path.exists(path): # move file before saving
-                    os.rename(path, tmp_path)
-                    #print(f"moved: {path} to {tmp_path}", flush=True)
-                if i == self.n_retry - 1:
-                    self._alternative_path = True
-                super()._save_model(epoch, batch, logs)
-                #print(f"saved epoch: {epoch}", flush=True)
+        if isinstance(self.save_freq, int) or self.epochs_since_last_save >= self.period:
+            for i in range(self.n_retry):
                 if i>1:
-                    print(f"model saved after retrying: {i+1}/{self.n_retry}", flush=True)
-            except BlockingIOError as error:
-                print(f"Error saving weights: {error}. Retry: {i+1}/{self.n_retry}", flush=True)
-                time.sleep(self.sleep_time)  # wait for X seconds before trying to save again
-                print(f"slept: {self.sleep_time}s")
-                os.remove(self._get_file_path(epoch, batch, logs))
-                print(f"file {self._get_file_path(epoch, batch, logs)} removed")
-                self._alternative_path = False
-            else:
-                self._alternative_path = False
-                if os.path.exists(tmp_path):
-                    os.remove(tmp_path)
-                return
-            print(f"save_model: end of for loop: {i+1}/{self.n_retry}")
+                    print(f"save_model: start of for loop: {i + 1}/{self.n_retry}")
+                try:
+                    self._alternative_path = False
+                    path = self._get_file_path(epoch, batch, logs)
+                    tmp_path = self._tmp_path(path)
+                    self._alternative_path = True
+                    super()._save_model(epoch, batch, logs) # save to tmp path
+                    self._alternative_path = False
+                    #print(f"save_model: try {i + 1}/{self.n_retry}, path: {path} temp path: {tmp_path}", flush=True)
+                    if os.path.exists(tmp_path):  # move tmp file to new file
+                        if os.path.exists(path):
+                            os.remove(path)
+                        os.rename(tmp_path, path)
+                        #print(f"moved: {path} to {tmp_path}", flush=True)
+                    if i == self.n_retry - 1:
+                        self._alternative_path = True
+                    super()._save_model(epoch, batch, logs)
+                    #print(f"saved epoch: {epoch}", flush=True)
+                except BlockingIOError as error:
+                    print(f"Error saving weights: {error}. Retry: {i+1}/{self.n_retry}", flush=True)
+                    time.sleep(self.sleep_time)  # wait for X seconds before trying to save again
+                    print(f"slept: {self.sleep_time}s")
+                    os.remove(self._get_file_path(epoch, batch, logs))
+                    print(f"file {self._get_file_path(epoch, batch, logs)} removed")
+                    self._alternative_path = False
+                else:
+                    self._alternative_path = False
+                    tmp = os.path.exists(tmp_path)
+                    main = os.path.exists(path)
+                    if main and tmp:
+                        os.remove(tmp_path)
+                        return
+                    elif main:
+                        return
+                    elif tmp and not main and i == self.n_retry-1:
+                        os.rename(tmp_path, path)
+                        print(f"model could not be saved at {self._get_file_path(epoch, batch, logs)} after {self.n_retry} retry, restoring temp file", flush=True)
+                        return
+                    time.sleep(self.sleep_time)
+                print(f"save_model: end of for loop: {i+1}/{self.n_retry}", flush=True)
 
         # 21:45:07.667:   File "h5py/h5f.pyx", line 126, in h5py.h5f.create
         # 21:45:07.667: BlockingIOError: [Errno 11] Unable to create file (unable to lock file, errno = 11, error message = 'Resource temporarily unavailable')
@@ -53,7 +67,7 @@ class SafeModelCheckpoint(ModelCheckpoint):
     def _get_file_path(self, epoch, batch, logs):
         path = super()._get_file_path(epoch, batch, logs)
         if self._alternative_path:
-            path = self._tmp_path(path)
+            return self._tmp_path(path)
         return path
 
     def _tmp_path(self, path):
